# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/objects": {
            "get": {
                "tags": [
                    "Objects"
                ],
                "summary": "Get all booking objects",
                "operationId": "2196cb81e8dc2947ce385a8e25bccb43",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "description": "Unique identifier for the booking object",
                                                "type": "integer"
                                            },
                                            "name": {
                                                "description": "Name of the booking object",
                                                "type": "string"
                                            },
                                            "description": {
                                                "description": "Description of the booking object",
                                                "type": "string"
                                            },
                                            "price": {
                                                "description": "Price of the booking object",
                                                "type": "integer"
                                            },
                                            "photos": {
                                                "description": "Photos for the booking object",
                                                "type": "string"
                                            },
                                            "preview_photo": {
                                                "description": "Preview photo of the booking object",
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Objects"
                ],
                "summary": "Create a new object",
                "operationId": "f6f7ca2320f4b9abc4afd95f20c91e80",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the booking object",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Description of the booking object",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Price of the booking object",
                                        "type": "integer"
                                    },
                                    "photos": {
                                        "description": "Photos for the booking object",
                                        "type": "string"
                                    },
                                    "preview_photo": {
                                        "description": "Preview photo of the object",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Object created successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
        "/api/objects/{id}": {
            "get": {
                "tags": [
                    "Objects"
                ],
                "summary": "Get a booking object by ID",
                "operationId": "68da07598cb1f5457fe1e2525050923a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the booking object",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "description": "Unique identifier for the booking object",
                                            "type": "integer"
                                        },
                                        "name": {
                                            "description": "Name of the booking object",
                                            "type": "string"
                                        },
                                        "description": {
                                            "description": "Description of the booking object",
                                            "type": "string"
                                        },
                                        "price": {
                                            "description": "Price of the booking object",
                                            "type": "integer"
                                        },
                                        "photos": {
                                            "description": "Photos for the booking object",
                                            "type": "string"
                                        },
                                        "preview_photo": {
                                            "description": "Preview photo of the booking object",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Object not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Objects"
                ],
                "summary": "Update a booking object by ID",
                "operationId": "4dcccf9ee985db5067baae7959de8bdd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the booking object",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Name of the booking object",
                                        "type": "string"
                                    },
                                    "description": {
                                        "description": "Description of the booking object",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Price of the booking object",
                                        "type": "integer"
                                    },
                                    "photos": {
                                        "description": "Photos for the booking object",
                                        "type": "string"
                                    },
                                    "preview_photo": {
                                        "description": "Preview photo of the booking object",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Object updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Success message",
                                            "type": "string"
                                        },
                                        "object": {
                                            "description": "Updated booking object details",
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Object not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Objects"
                ],
                "summary": "Delete a booking object by ID",
                "operationId": "1ebad6cbad824a3c8bd016fbfabb9960",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the booking object",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Object deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Success message",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Object not found"
                    }
                }
            }
        },
...
    },
...
}